+++ api/CHANGELOG.md
@@ -11,6 +11,9 @@ All notable changes to the **Prowler API** are documented in this file.
 - Add API scan report system, now all scans launched from the API will generate a compressed file with the report in OCSF, CSV and HTML formats [(#6878)](https://github.com/prowler-cloud/prowler/pull/6878).
 - Configurable Sentry integration [(#6874)](https://github.com/prowler-cloud/prowler/pull/6874)
 
+### Changed
+- Optimized `GET /findings` endpoint to improve response time and size [(#7019)](https://github.com/prowler-cloud/prowler/pull/7019).
+
 ---
 
 ## [v1.4.0] (Prowler v5.3.0) - 2025-02-10